:memo: Added CHANGELOG and UPGRADING information (#567)
This PR adds a dedicated changelog and upgrade guide to the project as a
centralized place to keep track of significant updates to this library.
Both of these files are curated selections of the PRs that go into a
release.
Notably, these are written for humans and not for machines.
As such, they are not automatically scraped together like the drafts
created by the release drafter action. These drafts may still serve as
blueprints for corresponding changelog entries.