✨  Implement HHL algorithm (#582)
Quite some time ago, the HHL implementation based on `qiskit.algorithms`
has been removed. Now, the time has come to re-add this benchmark.

In this current implementation, a simplified version is implemented that
solves a specific linear equation system and the `num_qubits` scale the
number of QPE qubits. This gives the advantages of getting a circuit
that actually has the specified number of qubits in comparison to
before.

However, while I think this implementation could and should be improved
in the future, this might be a fair starting point.

This PR resolves #216.

---------

Signed-off-by: Nils Quetschlich <nils.quetschlich@tum.de>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>