Modify the example (#689)
TypeError: get_benchmark() got an unexpected keyword argument
'benchmark_name'

## Description

Please include a summary of the change and, if applicable, which issue
is fixed.
Please also include relevant motivation and context.
List any dependencies that are required for this change.

Fixes #(issue) <!--- Replace `(issue)` with the issue number fixed by
this pull request. If this PR does not fix an issue, please remove this
line. -->

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [ ] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [ ] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [ ] I have added migration instructions to the upgrade guide (if
needed).
- [ ] The changes follow the project's style guidelines and introduce no
new warnings.
- [ ] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

Signed-off-by: Soroush Fathi <81929775+soroushfathi@users.noreply.github.com>